revamp calculations

+++ src/main/java/com/cob/feedback/model/performance/HospitalityContainer.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Getter
 @Builder
 public class HospitalityContainer {
-    private long happyIndex;
+    private double happyIndex;
     private long nps;
     private double average;
